Refactor fenwick tree
