+++ code/algorithms/data_structures/fenwick-tree-3d.cpp
@@ -32,9 +32,6 @@ struct FenwickTree {
     return ans;
   }
   Long Query(int x1, int y1, int z1, int x2, int y2, int z2) {
-    return Query(x2, y2, z2)
-      - Query(x1 - 1, y2, z2) - Query(x2, y1 - 1, z2) - Query(x2, y2, z1 - 1)
-      + Query(x1 - 1, y1 - 1, z2) + Query(x1 - 1, y2, z1 - 1)
-      + Query(x2, y1 - 1, z1 - 1) - Query(x1 - 1, y1 - 1, z1 - 1);
+    return Query(x2, y2, z2) - Query(x1 - 1, y2, z2) - Query(x2, y1 - 1, z2) - Query(x2, y2, z1 - 1) + Query(x1 - 1, y1 - 1, z2) + Query(x1 - 1, y2, z1 - 1) + Query(x2, y1 - 1, z1 - 1) - Query(x1 - 1, y1 - 1, z1 - 1);
   }
 } ft;
