+++ code/tests/data-structures/fenwick-tree-3d-test.cpp
@@ -1,4 +1,4 @@
-//https://github.com/Sky-Nik/Codeforces/blob/main/FenwickTree/statements.pdf
+// https://github.com/Sky-Nik/Codeforces/blob/main/FenwickTree/statements.pdf
 #include <bits/stdc++.h>
 using namespace std;
 
@@ -39,10 +39,7 @@ struct FenwickTree {
     return ans;
   }
   Long Query(int x1, int y1, int z1, int x2, int y2, int z2) {
-    return Query(x2, y2, z2)
-      - Query(x1 - 1, y2, z2) - Query(x2, y1 - 1, z2) - Query(x2, y2, z1 - 1)
-      + Query(x1 - 1, y1 - 1, z2) + Query(x1 - 1, y2, z1 - 1)
-      + Query(x2, y1 - 1, z1 - 1) - Query(x1 - 1, y1 - 1, z1 - 1);
+    return Query(x2, y2, z2) - Query(x1 - 1, y2, z2) - Query(x2, y1 - 1, z2) - Query(x2, y2, z1 - 1) + Query(x1 - 1, y1 - 1, z2) + Query(x1 - 1, y2, z1 - 1) + Query(x2, y1 - 1, z1 - 1) - Query(x1 - 1, y1 - 1, z1 - 1);
   }
 } ft;
